moved multiple parserResult definitions to one file
